+++ Sources/src-tauri/engine/SmartSwitchKey.cpp
@@ -54,7 +54,7 @@ void getSmartSwitchKeySaveData(vector<Byte>& outData) {
     
     for (std::map<string, Int8>::iterator it = _smartSwitchKeyData.begin(); it != _smartSwitchKeyData.end(); ++it) {
         outData.push_back((Byte)it->first.length());
-        for (int j = 0; j < it->first.length(); j++) {
+        for (size_t j = 0; j < it->first.length(); j++) {
             outData.push_back(it->first[j]);
         }
         outData.push_back(it->second);
